update store to get fetch funcitons in one file
